feat(daemon): embeddable mode behind PITCHBOX_EMBED_DAEMON=1
Extracts the daemon loops into startEmbeddedDaemon() so the web server
can run them in-proc on single-host installs, removing the need to
supervise a second Node process. Standalone daemon remains the default
deploy for HA / multi-host setups.

The dispatch lock (#32) and webhook DLQ SKIP LOCKED (#36) keep the
guarantees intact when both modes run against the same DB. Heartbeats
from embedded mode use module='web' so Settings can tell them apart.
