+++ daemon/src/index.ts
@@ -1,151 +1,22 @@
-import { applyJitter } from '@pitchbox/shared/scheduler/jitter';
-import { config } from './config.js';
+/**
+ * Standalone daemon entry point. Wraps the embeddable factory with process
+ * signal handlers so SIGINT/SIGTERM drain in-flight iterations before exit.
+ */
 import { logger } from './logger.js';
-import { beat } from './heartbeat.js';
-import { tick as schedulerTick } from './scheduler.js';
-import { tick as replyPollerTick } from './reply-poller.js';
-import { tick as webhookSenderTick } from './webhook-sender.js';
-import { tick as retentionTick } from './retention.js';
-import { tick as keywordWatcherTick } from './keyword-watcher.js';
+import { startEmbeddedDaemon } from './embed.js';
 
 const log = logger('main');
 
-let running = true;
+const daemon = startEmbeddedDaemon({ heartbeatModule: 'daemon' });
 
-type Loop = {
-  name: string;
-  intervalMs: number;
-  run: () => Promise<void>;
+let shuttingDown = false;
+const shutdown = async (sig: string) => {
+  if (shuttingDown) return;
+  shuttingDown = true;
+  log.info(`received ${sig}`);
+  await daemon.stop();
+  process.exit(0);
 };
 
-type LoopHandle = { cancel: () => void; settle: () => Promise<void> };
-
-function every(ms: number, fn: () => Promise<void>): LoopHandle {
-  let timer: ReturnType<typeof setTimeout> | null = null;
-  let cancelled = false;
-  let inflight: Promise<void> | null = null;
-
-  const kick = async () => {
-    if (cancelled) return;
-    const run = (async () => {
-      try {
-        await fn();
-      } catch (err) {
-        log.error('loop iteration crashed', err);
-      }
-    })();
-    inflight = run;
-    try {
-      await run;
-    } finally {
-      inflight = null;
-      // Apply symmetric jitter so concurrent loops (and multi-instance daemons)
-      // don't lock-step on the same tick.
-      if (!cancelled) timer = setTimeout(kick, applyJitter(ms, config.jitterPct));
-    }
-  };
-
-  void kick();
-  return {
-    cancel() {
-      cancelled = true;
-      if (timer) clearTimeout(timer);
-    },
-    async settle() {
-      if (inflight) await inflight;
-    },
-  };
-}
-
-async function main() {
-  log.info(`pitchbox daemon starting (web=${config.webUrl})`);
-
-  const loops: Loop[] = [
-    {
-      name: 'heartbeat',
-      intervalMs: config.heartbeatIntervalMs,
-      run: () => beat('daemon'),
-    },
-    {
-      name: 'scheduler',
-      intervalMs: config.tickIntervalMs,
-      run: schedulerTick,
-    },
-  ];
-
-  loops.push({
-    name: 'retention',
-    intervalMs: config.retentionIntervalMs,
-    run: async () => {
-      const res = await retentionTick();
-      if (res.runEventsDeleted + res.draftEventsDeleted + res.draftsDeleted > 0) {
-        logger('retention').info(
-          `pruned run_events=${res.runEventsDeleted} draft_events=${res.draftEventsDeleted} drafts=${res.draftsDeleted}`,
-        );
-      }
-    },
-  });
-
-  loops.push({
-    name: 'keyword-watcher',
-    intervalMs: config.keywordWatcherIntervalMs,
-    run: async () => {
-      const res = await keywordWatcherTick();
-      if (res.checked > 0) {
-        logger('keyword-watcher').info(
-          `checked ${res.checked} watches → ${res.dispatched} dispatched`,
-        );
-      }
-    },
-  });
-
-  loops.push({
-    name: 'webhook-sender',
-    intervalMs: config.webhookSenderIntervalMs,
-    run: async () => {
-      await webhookSenderTick();
-    },
-  });
-
-  if (!config.repliesDisabled) {
-    loops.push({
-      name: 'reply-poller',
-      intervalMs: config.replyPollIntervalMs,
-      run: async () => {
-        const res = await replyPollerTick();
-        if (res.checked > 0) {
-          logger('reply-poller').info(
-            `checked ${res.checked} contacts → ${res.newReplies} new replies, ${res.skipped} skipped`,
-          );
-        }
-      },
-    });
-  }
-
-  const cancels = loops.map((l) => {
-    log.info(`starting loop "${l.name}" every ${l.intervalMs}ms`);
-    return every(l.intervalMs, l.run);
-  });
-
-  const shutdown = async (sig: string) => {
-    if (!running) return;
-    running = false;
-    log.info(`received ${sig}, draining loops`);
-    for (const c of cancels) c.cancel();
-    // Wait up to 10s for in-flight iterations (HTTP POST to /api/run, reply
-    // polling) to finish — beats killing them mid-request.
-    const drain = Promise.all(cancels.map((c) => c.settle()));
-    const timeout = new Promise<void>((resolve) => setTimeout(resolve, 10_000));
-    await Promise.race([drain, timeout]);
-    log.info('exit');
-    process.exit(0);
-  };
-
-  process.on('SIGINT', () => void shutdown('SIGINT'));
-  process.on('SIGTERM', () => void shutdown('SIGTERM'));
-}
-
-main().catch((err) => {
-  log.error('fatal', err);
-  process.exit(1);
-});
+process.on('SIGINT', () => void shutdown('SIGINT'));
+process.on('SIGTERM', () => void shutdown('SIGTERM'));
