+++ web/package.json
@@ -14,6 +14,7 @@
   },
   "dependencies": {
     "@bytemd/plugin-gfm": "^1.22.0",
+    "@pitchbox/daemon": "*",
     "@pitchbox/shared": "*",
     "@tailwindcss/vite": "^4.2.4",
     "@types/dompurify": "^3.0.5",
